refactor: extract Logo component out of Header
